move tests to main library dir
