+++ adie/tests/data/bids_test/sourcedata/mytask/edge_case/level2/ADIE983F/data.csv
@@ -1,0 +1,5 @@
+stim,resp,rt,id,ses
+test,1,0.042,983,01
+test,1,0.042,983,01
+test,1,0.042,983,01
+test,1,0.042,983,01
