+++ adie/tests/data/bids_test/sub-ADIE983_ses-baseline_task-mytask_beh.tsv
@@ -1,0 +1,5 @@
+stim,resp,rt,id,ses
+test,1,0.042,983,01
+test,1,0.042,983,01
+test,1,0.042,983,01
+test,1,0.042,983,01
