+++ adie/tests/data/bids_test/sub-ADIE999/ses-3mf/beh/sub-ADIE999_ses-3mf_task-mytask_beh.tsv
@@ -1,0 +1,5 @@
+stim	resp	rt	id	ses
+test	1	0.042	983	1
+test	1	0.042	983	1
+test	1	0.042	983	1
+test	1	0.042	983	1
