+++ adie/tests/data/bids_test/sub-CONADIE999/ses-baseline/beh/sub-CONADIE999_ses-baseline_task-mytask_stim.json
@@ -1,0 +1,10 @@
+{
+  "SamplingFrequency": 1000.0,
+  "StartTime": 0.0,
+  "Columns": [
+    "stim"
+  ],
+  "Stim": {
+    "Unit": "arbitrary"
+  }
+}
