--- test/data/bids_test/sourcedata/mytask/level1/anotherlevel/ADIE999_BL/data.csv
@@ -1,5 +1,0 @@
-stim,resp,rt,id,ses
-test,1,0.042,983,01
-test,1,0.042,983,01
-test,1,0.042,983,01
-test,1,0.042,983,01
