--- test/data/bids_test/sub-ADIE983/beh/sub-ADIE983_ses-baseline_task-mytask_beh.tsv
@@ -1,5 +1,0 @@
-stim,resp,rt,id,ses
-test,1,0.042,983,01
-test,1,0.042,983,01
-test,1,0.042,983,01
-test,1,0.042,983,01
